JBPM-8067: Stunner - Magnets & connector bendpoints are not working with the growable canvas. (#2356)

+++ kie-wb-common-stunner/kie-wb-common-stunner-client/kie-wb-common-stunner-widgets/src/main/java/org/kie/workbench/common/stunner/client/widgets/canvas/StunnerBoundsProviderFactory.java
@@ -21,7 +21,7 @@ import static com.ait.lienzo.client.widget.panel.impl.BoundsProviderFactory.comp
 
 public class StunnerBoundsProviderFactory {
 
-    public static final double ASPECT_RATIO = 300 / 150;
+    public static final float ASPECT_RATIO = 300 / 150;
     public static final double PADDING = 50d;
 
     public static WiresBoundsProvider newProvider() {
